Compass display cleaned up. CustomMapView moved to components-directory and renderDegreeMarkres moved to utils directory. Utils-rirectory initialized.
